TUNG PHAN: IMPLEMENT SQLITE DATABASE
- Implement sqlite database using SQLiteHelper
- Save assets file to external storage only one time.
- Get survey report from database, insert or get patient survey is allowed.
- Put field id in some json object to make it unique.

+++ app/src/main/java/com/wizeline/tungphan/wizelinealzheirmersurvey/WizeApp.java
@@ -77,8 +77,12 @@ public class WizeApp extends Application {
         try {
             in = assetManager.open(filename);
             File outFile = new File(getExternalFilesDir(null), filename);
-            out = new FileOutputStream(outFile);
-            copyFile(in, out);
+            if(!outFile.exists()) {
+                out = new FileOutputStream(outFile);
+                copyFile(in, out);
+            }else{
+                Log.e(TAG,"files copied");
+            }
         } catch (IOException e) {
             Log.e(TAG, e.getMessage());
         } finally {
